# helm/usdsearch/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "USD Search and Asset Graph Search APIs",
...
    "version": "1.4.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/llm_parse/query": {
      "post": {
        "tags": [
          "AI Search"
        ],
        "summary": "Parse a free-text query into structured search parameters",
        "description": "Uses an LLM with constrained (JSON-schema) decoding to turn free-text into a structured intermediate representation, then deterministically maps it onto search request fields. Returns a 503 if the feature is disabled on this server or the LLM call fails.",
        "operationId": "llm_parse_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryParsingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryParsingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "DroppedFilter": {
        "properties": {
          "filter": {
            "$ref": "#/components/schemas/IRFilter",
            "description": "The extracted filter that was not applied."
          },
          "reason": {
            "$ref": "#/components/schemas/DroppedFilterReason",
            "description": "Machine-readable reason the filter was dropped."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable explanation suitable for surfacing in the UI."
          }
        },
        "type": "object",
        "required": [
          "filter",
          "reason",
          "message"
        ],
        "title": "DroppedFilter",
        "description": "A filter the LLM extracted that the mapper could not apply, with the\nreason. Lets the client tell the user a constraint was understood but not\nenforced (and why), instead of silently ignoring it."
      },
      "DroppedFilterReason": {
        "type": "string",
        "enum": [
          "unknown_field",
          "unsupported_operator",
          "invalid_value"
        ],
        "title": "DroppedFilterReason",
        "description": "Why the mapper could not apply an extracted filter."
      },
...
      "IRFieldName": {
        "type": "string",
        "enum": [
          "file_type",
          "name",
          "path_prefix",
          "created_at",
          "modified_at",
          "size",
          "created_by",
          "tag",
          "usd_property",
          "polygon_count",
          "dimension"
        ],
        "title": "IRFieldName"
      },
      "IRFilter": {
        "properties": {
          "field": {
            "$ref": "#/components/schemas/IRFieldName",
            "description": "The field to filter on (one of the allowed enum values)."
          },
          "operator": {
            "$ref": "#/components/schemas/Operator",
            "description": "Comparison operator. Must be valid for the chosen field."
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "title": "Value",
            "description": "The value to compare against. Use a list only for `in` operators."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "operator",
          "value"
        ],
        "title": "IRFilter",
        "description": "A single structured filter: which field, which operator, what value."
      },
...
      "Operator": {
        "type": "string",
        "enum": [
          "eq",
          "neq",
          "gt",
          "gte",
          "lt",
          "lte",
          "contains",
          "in",
          "exists"
        ],
        "title": "Operator",
        "description": "Comparison operators an IR filter may use. The set is the union across\nall fields; each field constrains which subset is valid (see ``IRField``)."
      },
...
      "QueryParsingRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query",
            "description": "Raw free-text search text from the search bar."
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "QueryParsingRequest"
      },
      "QueryParsingResponse": {
        "properties": {
          "interpreted_query": {
            "$ref": "#/components/schemas/SearchIR",
            "description": "Structured interpretation of the query."
          },
          "search_params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Search Params",
            "description": "Mapped DeepSearchSearchRequestV2 fields that were merged into the search request."
          },
          "dropped_filters": {
            "items": {
              "$ref": "#/components/schemas/DroppedFilter"
            },
            "type": "array",
            "title": "Dropped Filters",
            "description": "Filters that were extracted but could not be applied (unsupported operator, unknown field, or invalid value). Empty when every extracted filter was applied; its length is the dropped count. interpreted_query reflects only applied filters."
          },
          "unmapped_constraints": {
            "items": {
              "$ref": "#/components/schemas/UnmappedConstraint"
            },
            "type": "array",
            "title": "Unmapped Constraints",
            "description": "Constraints the user expressed that match no catalog field, so they could not become a filter and were not searched on. Surfaced (verbatim, with a note) so the client can tell the user the constraint was understood but not enforced. Mirrors interpreted_query.unmapped_constraints."
          }
        },
        "type": "object",
        "required": [
          "interpreted_query",
          "search_params"
        ],
        "title": "QueryParsingResponse",
        "description": "Same shape as ParsedQuery; kept as a distinct schema name for\n/llm_parse/query back-compat."
      },
...
      "SearchIR": {
        "properties": {
          "semantic_query": {
            "type": "string",
            "title": "Semantic Query",
            "description": "Descriptive nouns/phrase for semantic + keyword search (e.g. 'warehouse'). Put ONLY descriptive nouns/adjectives here \u2014 not structured constraints. May be empty if the query is purely a filter."
          },
          "filters": {
            "items": {
              "$ref": "#/components/schemas/IRFilter"
            },
            "type": "array",
            "title": "Filters",
            "description": "Structured filters extracted from the query. Empty if the query maps to no known filter."
          },
          "unmapped_constraints": {
            "items": {
              "$ref": "#/components/schemas/UnmappedConstraint"
            },
            "type": "array",
            "title": "Unmapped Constraints",
            "description": "Constraints the user clearly expressed that match NO catalog field. Never drop such a clause and never put it in semantic_query \u2014 record it here so the client can report it as understood-but-not-enforced."
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit",
            "description": "Maximum number of results, only if the user explicitly asked for a count."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "semantic_query"
        ],
        "title": "SearchIR",
        "description": "Structured interpretation of a free-text search query."
      },
...
      "UnmappedConstraint": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "The verbatim fragment of the user's query expressing this constraint."
          },
          "note": {
            "type": "string",
            "title": "Note",
            "description": "Plain-English explanation of what was understood and why no catalog field fits."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "note"
        ],
        "title": "UnmappedConstraint",
        "description": "A constraint the user clearly expressed that maps to no catalog field.\n\nThe escape hatch that keeps the parse honest: a clause like \"Isaac version\nbetween 5 and 6\" is neither a descriptive noun (so it does not belong in\n``semantic_query``) nor a known filter (so it cannot become an ``IRFilter``).\nRouting it here makes the drop explicit \u2014 the client can tell the user the\nconstraint was understood but not enforced \u2014 instead of it silently vanishing\nor polluting the semantic query."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
...
}